parallel-import: move format_stdout in its own file
