+++ import_data/tasks/tasks.py
@@ -10,18 +10,18 @@ import os
 import os.path
 import time
 from datetime import timedelta
-from functools import wraps
 import requests
-from invoke import context, task
+from invoke import task
 from invoke.exceptions import Failure
 import osmium
 
 from .lock import FileLock
 from .download import needs_to_download, download_file
+from .format_stdout import format_stdout
 
 cc_exec = concurrent.futures.ThreadPoolExecutor()
 
-def join_futures(self, *futures):
+def join_futures(*futures):
     """
     Wait for the result of all input futures and returns the list of results.
     Throws an exception if any of the futures fails.
@@ -30,59 +30,6 @@ def join_futures(self, *futures):
 
 logging.basicConfig(level=logging.INFO)
 
-
-class PrefixedStream:
-    """
-    Wrapper class around a stream adding a prefix to each writen line.
-    """
-    def __init__(self, src_stream, prefix):
-        self.prefix = prefix
-        self.src_stream = src_stream
-        self.buffer = ''
-
-    def write(self, data):
-        self.buffer += data
-        self.flush()
-
-    def flush(self):
-        lines = self.buffer.split('\n')
-
-        for line in lines[:-1]:
-            self.src_stream.write(
-                '[{}] {}\n'.format(self.prefix, line)
-            )
-
-        self.buffer = lines[-1]
-
-
-class PrefixedContext(context.Context):
-    """
-    Wrapper class around a pyinvoke context adding a prefix to each lines
-    outputed by ctx.run(..).
-    """
-    def __init__(self, ctx, prefix):
-        super().__init__(ctx.config)
-        self.prefix = prefix
-
-    def run(self, *args, **kwargs):
-        return super().run(
-            *args,
-            **kwargs,
-            out_stream=PrefixedStream(sys.stdout, self.prefix),
-            err_stream=PrefixedStream(sys.stderr, self.prefix + ':ERR'),
-        )
-
-def format_stdout(fun):
-    """
-    Wrapper decorator around a task adding its name to each line outputed by
-    ctx.run(..).
-    """
-    @wraps(fun)
-    def upgraded_fun(ctx, *args, **kwargs):
-        ctx = PrefixedContext(ctx, fun.__name__)
-        fun(ctx, *args, **kwargs)
-
-    return upgraded_fun
 
 class TilesLayer:
     BASEMAP = 'basemap'
